% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxode2lincmt-package.R
\docType{package}
\name{rxode2lincmt-package}
\alias{rxode2lincmt}
\alias{rxode2lincmt-package}
\title{rxode2lincmt: Linear Compartment Model Solutions and Gradients for 'rxode2'}
\description{
Analytic one, two and three compartment linear pharmacokinetic solutions with their parameter gradients from 'stan' automatic differentiation (Carpenter et al (2015) \doi{10.48550/arXiv.1509.07164}), eigen decompositions and derived-parameter conversions used by 'rxode2' (Wang, Hallow and James (2016) \doi{10.1002/psp4.12052}). Split out of 'rxode2' so its installation does not compile 'stan' AD items which made it take too long to compile by itself. The closed-form solutions follow the idea of the 'wnl' package (Bae, \url{https://CRAN.R-project.org/package=wnl}), though the implementation here is different.
}
\details{
The closed-form linear compartment solutions follow the idea of
the 'wnl' package by Kyun-Seop Bae
(\url{https://CRAN.R-project.org/package=wnl}); the implementation here is
different.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nlmixr2.github.io/rxode2lincmt/}
  \item \url{https://github.com/nlmixr2/rxode2lincmt/}
  \item Report bugs at \url{https://github.com/nlmixr2/rxode2lincmt/issues/}
}

}
\author{
\strong{Maintainer}: Matthew L. Fidler \email{matthew.fidler@gmail.com} (\href{https://orcid.org/0000-0001-8538-6691}{ORCID})

Authors:
\itemize{
  \item Matthew L. Fidler \email{matthew.fidler@gmail.com} (\href{https://orcid.org/0000-0001-8538-6691}{ORCID})
}

Other contributors:
\itemize{
  \item Richard Upton [contributor]
}

}
\keyword{internal}
