% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solComp.R
\name{solComp2}
\alias{solComp2}
\title{Eigenvalues and coefficients of the two compartment model}
\usage{
solComp2(k10, k12, k21)
}
\arguments{
\item{k10}{elimination rate}

\item{k12}{rate from central to peripheral compartment}

\item{k21}{rate from peripheral to central compartment}
}
\value{
List with \code{L} vector and matrices \code{C1} and \code{C2}
}
\description{
Computes the exponents (\code{L}) and coefficient matrices (\code{C1}, \code{C2}) of the
two compartment linear system from its micro-constants.  The idea comes
from the 'wnl' package by Kyun-Seop Bae; the implementation is different.
}
\examples{
solComp2(k10 = 0.1, k12 = 3, k21 = 1)
}
\author{
Matthew L. Fidler, based on the idea in the 'wnl' package
}
