% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solComp.R
\name{solComp3}
\alias{solComp3}
\title{Eigenvalues and coefficients of the three compartment model}
\usage{
solComp3(k10, k12, k21, k13, k31)
}
\arguments{
\item{k10}{elimination rate}

\item{k12}{rate from central to peripheral compartment}

\item{k21}{rate from peripheral to central compartment}

\item{k13}{rate from central to peripheral compartment #2}

\item{k31}{rate from peripheral compartment #2 to central}
}
\value{
List with \code{L} vector and matrices \code{C1}, \code{C2} and \code{C3}
}
\description{
Eigenvalues and coefficients of the three compartment model
}
\examples{
solComp3(k10 = 0.1, k12 = 3, k21 = 1, k13 = 2, k31 = 0.5)
}
\author{
Matthew L. Fidler
}
